right side bar

+++ routes/admin/purchases/index.js
@@ -229,8 +229,8 @@ router.post('/add', PermissionModule.Permission('add', moduleSlug,  extraVar), f
                             item_id: value1.item_id,
                             type: "purchase",
                             sub_item_id: value1.sub_item_id,
-                            quantity: parseInt(value1.quantity) + parseInt(data1.quantity),
-                            no_of_pkg: parseInt(value1.no_of_pkg) + parseInt(data1.no_of_pkg),
+                            quantity: parseFloat(value1.quantity) + parseFloat(data1.quantity),
+                            no_of_pkg: parseFloat(value1.no_of_pkg) + parseFloat(data1.no_of_pkg),
                           };
                           models.Stock.updateAllValues(stockDataUpdate, function (results) {
                             callback1();
@@ -266,8 +266,8 @@ router.post('/add', PermissionModule.Permission('add', moduleSlug,  extraVar), f
                             item_id: value1.item_id,
                             type: "purchase",
                             sub_item_id: value1.sub_item_id !== "" ? value1.sub_item_id : null,
-                            quantity: parseInt(value1.quantity) + parseInt(data.quantity),
-                            no_of_pkg: parseInt(value1.no_of_pkg) + parseInt(data.no_of_pkg),
+                            quantity: parseFloat(value1.quantity) + parseFloat(data.quantity),
+                            no_of_pkg: parseFloat(value1.no_of_pkg) + parseFloat(data.no_of_pkg),
                           };
                           models.Stock.updateAllValues(stockDataUpdate, function (results) {
                             callback1();
@@ -578,15 +578,15 @@ router.post('/edit', PermissionModule.Permission('edit', moduleSlug,  extraVar),
                           if(previousPurchaseItemValue[data1.sub_item_id]){
                             console.log("sub exist22222222")
                             let tmpPre = parseInt(previousPurchaseItemValue[data1.sub_item_id]);
-                            let tmpPost = parseInt(value1.quantity);
+                            let tmpPost = parseFloat(value1.quantity);
                             let qty = 0;
                             
                             if(tmpPre > tmpPost){
-                              qty = parseInt(data1.quantity) - parseInt(tmpPre - tmpPost);
+                              qty = parseFloat(data1.quantity) - parseFloat(tmpPre - tmpPost);
                             } else if(tmpPre < tmpPost){
-                              qty = parseInt(data1.quantity) + parseInt(tmpPost - tmpPre);
+                              qty = parseFloat(data1.quantity) + parseFloat(tmpPost - tmpPre);
                             } else {
-                              qty = parseInt(data1.quantity);
+                              qty = parseFloat(data1.quantity);
                             }
                             stockDataUpdate.body = {
                               id: data1.id,
@@ -594,7 +594,7 @@ router.post('/edit', PermissionModule.Permission('edit', moduleSlug,  extraVar),
                               type: "purchase",
                               sub_item_id: value1.sub_item_id,
                               quantity: qty,
-                              no_of_pkg: parseInt(value1.no_of_pkg) + parseInt(data1.no_of_pkg),
+                              no_of_pkg: parseFloat(value1.no_of_pkg) + parseFloat(data1.no_of_pkg),
                             };
                             console.log("sub exist22222222", stockDataUpdate)
                           } else {
@@ -603,8 +603,8 @@ router.post('/edit', PermissionModule.Permission('edit', moduleSlug,  extraVar),
                               item_id: value1.item_id,
                               type: "purchase",
                               sub_item_id: value1.sub_item_id,
-                              quantity: parseInt(value1.quantity) + parseInt(data1.quantity),
-                              no_of_pkg: parseInt(value1.no_of_pkg) + parseInt(data1.no_of_pkg),
+                              quantity: parseFloat(value1.quantity) + parseFloat(data1.quantity),
+                              no_of_pkg: parseFloat(value1.no_of_pkg) + parseFloat(data1.no_of_pkg),
                             };
                           }
                           console.log("sub exist3333", stockDataUpdate)
@@ -639,15 +639,15 @@ router.post('/edit', PermissionModule.Permission('edit', moduleSlug,  extraVar),
                           if(previousPurchaseItemValue[data.item_id]){
                             console.log("&&&1");
                             let tmpPre = parseInt(previousPurchaseItemValue[data.item_id]);
-                            let tmpPost = parseInt(value1.quantity);
+                            let tmpPost = parseFloat(value1.quantity);
                             let qty = 0;
                             console.log("&&&1", tmpPre , tmpPost);
                             if(tmpPre > tmpPost){
-                              qty = parseInt(data.quantity) - parseInt(tmpPre - tmpPost);
+                              qty = parseFloat(data.quantity) - parseFloat(tmpPre - tmpPost);
                             } else if(tmpPre < tmpPost){
-                              qty = parseInt(data.quantity) + parseInt(tmpPost - tmpPre);
+                              qty = parseFloat(data.quantity) + parseFloat(tmpPost - tmpPre);
                             } else {
-                              qty = parseInt(data.quantity);
+                              qty = parseFloat(data.quantity);
                             }
 
                             stockDataUpdate.body = {
@@ -656,7 +656,7 @@ router.post('/edit', PermissionModule.Permission('edit', moduleSlug,  extraVar),
                               type: "purchase",
                               sub_item_id: value1.sub_item_id !== "" ? value1.sub_item_id : null,
                               quantity: qty,
-                              no_of_pkg: parseInt(value1.no_of_pkg) + parseInt(data.no_of_pkg),
+                              no_of_pkg: parseFloat(value1.no_of_pkg) + parseFloat(data.no_of_pkg),
                             };
                             console.log('exist', stockDataUpdate);
                           } else {
@@ -667,8 +667,8 @@ router.post('/edit', PermissionModule.Permission('edit', moduleSlug,  extraVar),
                               item_id: value1.item_id,
                               type: "purchase",
                               sub_item_id: value1.sub_item_id !== "" ? value1.sub_item_id : null,
-                              quantity: parseInt(value1.quantity) + parseInt(data.quantity),
-                              no_of_pkg: parseInt(value1.no_of_pkg) + parseInt(data.no_of_pkg),
+                              quantity: parseFloat(value1.quantity) + parseFloat(data.quantity),
+                              no_of_pkg: parseFloat(value1.no_of_pkg) + parseFloat(data.no_of_pkg),
                             };
                           }
                           console.log("########", stockDataUpdate)
