+++ routes/admin/stocks/index.js
@@ -62,50 +62,50 @@ router.get('/stock_movement', PermissionModule.Permission('view', moduleSlug,  e
         results.map(function(val, index){
           if(val.sub_item_id){
             if(purchaseItemsObj[val.sub_item_id+'_purchase']){
-              purchaseItemsObj[val.sub_item_id+'_purchase']['quantity'] = parseInt(purchaseItemsObj[val.sub_item_id+'_purchase']['quantity']) + parseInt(val.quantity);
+              purchaseItemsObj[val.sub_item_id+'_purchase']['quantity'] = parseFloat(purchaseItemsObj[val.sub_item_id+'_purchase']['quantity']) + parseFloat(val.quantity);
               purchaseItemsObj[val.sub_item_id+'_purchase'] = purchaseItemsObj[val.sub_item_id+'_purchase'];
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                purchaseItemsObj[val.sub_item_id+'_purchase']['today'] = parseInt(purchaseItemsObj[val.sub_item_id+'_purchase']['today']) + parseInt(val.quantity);
+                purchaseItemsObj[val.sub_item_id+'_purchase']['today'] = parseFloat(purchaseItemsObj[val.sub_item_id+'_purchase']['today']) + parseFloat(val.quantity);
               } else {
-                purchaseItemsObj[val.sub_item_id+'_purchase']['old'] = parseInt(purchaseItemsObj[val.sub_item_id+'_purchase']['old']) + parseInt(val.quantity);
+                purchaseItemsObj[val.sub_item_id+'_purchase']['old'] = parseFloat(purchaseItemsObj[val.sub_item_id+'_purchase']['old']) + parseFloat(val.quantity);
               }
-              // purchaseItemsObj[val.sub_item_id+'_purchase'] = parseInt(purchaseItemsObj[val.sub_item_id+'_purchase']) + parseInt(val.quantity);
+              // purchaseItemsObj[val.sub_item_id+'_purchase'] = parseFloat(purchaseItemsObj[val.sub_item_id+'_purchase']) + parseFloat(val.quantity);
             } else {
               purchaseItemsObj[val.sub_item_id+'_purchase'] = {};
-              purchaseItemsObj[val.sub_item_id+'_purchase']['quantity'] = parseInt(val.quantity);
+              purchaseItemsObj[val.sub_item_id+'_purchase']['quantity'] = parseFloat(val.quantity);
               purchaseItemsObj[val.sub_item_id+'_purchase']['today'] = 0;
               purchaseItemsObj[val.sub_item_id+'_purchase']['old'] = 0;
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                purchaseItemsObj[val.sub_item_id+'_purchase']['today'] = parseInt(purchaseItemsObj[val.sub_item_id+'_purchase']['today']) + parseInt(val.quantity);
+                purchaseItemsObj[val.sub_item_id+'_purchase']['today'] = parseFloat(purchaseItemsObj[val.sub_item_id+'_purchase']['today']) + parseFloat(val.quantity);
               } else {
-                purchaseItemsObj[val.sub_item_id+'_purchase']['old'] = parseInt(purchaseItemsObj[val.sub_item_id+'_purchase']['old']) + parseInt(val.quantity);
+                purchaseItemsObj[val.sub_item_id+'_purchase']['old'] = parseFloat(purchaseItemsObj[val.sub_item_id+'_purchase']['old']) + parseFloat(val.quantity);
               }
               
             }
           } else {
             if(purchaseItemsObj[val.item_id+'_purchase']){
-              purchaseItemsObj[val.item_id+'_purchase']['quantity'] = parseInt(purchaseItemsObj[val.item_id+'_purchase']['quantity']) + parseInt(val.quantity);
+              purchaseItemsObj[val.item_id+'_purchase']['quantity'] = parseFloat(purchaseItemsObj[val.item_id+'_purchase']['quantity']) + parseFloat(val.quantity);
               purchaseItemsObj[val.item_id+'_purchase'] = purchaseItemsObj[val.item_id+'_purchase'];
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                purchaseItemsObj[val.item_id+'_purchase']['today'] = parseInt(purchaseItemsObj[val.item_id+'_purchase']['today']) + parseInt(val.quantity);
+                purchaseItemsObj[val.item_id+'_purchase']['today'] = parseFloat(purchaseItemsObj[val.item_id+'_purchase']['today']) + parseFloat(val.quantity);
               } else {
-                purchaseItemsObj[val.item_id+'_purchase']['old'] = parseInt(purchaseItemsObj[val.item_id+'_purchase']['old']) + parseInt(val.quantity);
+                purchaseItemsObj[val.item_id+'_purchase']['old'] = parseFloat(purchaseItemsObj[val.item_id+'_purchase']['old']) + parseFloat(val.quantity);
               }
-              // purchaseItemsObj[val.item_id+'_purchase'] = parseInt(purchaseItemsObj[val.item_id+'_purchase']) + parseInt(val.quantity);
+              // purchaseItemsObj[val.item_id+'_purchase'] = parseFloat(purchaseItemsObj[val.item_id+'_purchase']) + parseFloat(val.quantity);
             } else {
               purchaseItemsObj[val.item_id+'_purchase'] = {};
-              purchaseItemsObj[val.item_id+'_purchase']['quantity'] = parseInt(val.quantity);
+              purchaseItemsObj[val.item_id+'_purchase']['quantity'] = parseFloat(val.quantity);
               purchaseItemsObj[val.item_id+'_purchase']['today'] = 0;
               purchaseItemsObj[val.item_id+'_purchase']['old'] = 0;
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
                 console.log(val.item_id+'_purchase', val.quantity)
-                purchaseItemsObj[val.item_id+'_purchase']['today'] = parseInt(purchaseItemsObj[val.item_id+'_purchase']['today']) + parseInt(val.quantity);
+                purchaseItemsObj[val.item_id+'_purchase']['today'] = parseFloat(purchaseItemsObj[val.item_id+'_purchase']['today']) + parseFloat(val.quantity);
               } else {
-                purchaseItemsObj[val.item_id+'_purchase']['old'] = parseInt(purchaseItemsObj[val.item_id+'_purchase']['old']) + parseInt(val.quantity);
+                purchaseItemsObj[val.item_id+'_purchase']['old'] = parseFloat(purchaseItemsObj[val.item_id+'_purchase']['old']) + parseFloat(val.quantity);
               }
             }
           }
@@ -121,48 +121,48 @@ router.get('/stock_movement', PermissionModule.Permission('view', moduleSlug,  e
         results.map(function(val, index){
           if(val.sub_item_id){
             if(productionItemsObj[val.sub_item_id+'_production']){
-              productionItemsObj[val.sub_item_id+'_production']['quantity'] = parseInt(productionItemsObj[val.sub_item_id+'_production']['quantity']) + parseInt(val.quantity);
+              productionItemsObj[val.sub_item_id+'_production']['quantity'] = parseFloat(productionItemsObj[val.sub_item_id+'_production']['quantity']) + parseFloat(val.quantity);
               productionItemsObj[val.sub_item_id+'_production'] = productionItemsObj[val.sub_item_id+'_production'];
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                productionItemsObj[val.sub_item_id+'_production']['today'] = parseInt(productionItemsObj[val.sub_item_id+'_production']['today']) + parseInt(val.quantity);
+                productionItemsObj[val.sub_item_id+'_production']['today'] = parseFloat(productionItemsObj[val.sub_item_id+'_production']['today']) + parseFloat(val.quantity);
               } else {
-                productionItemsObj[val.sub_item_id+'_production']['old'] = parseInt(productionItemsObj[val.sub_item_id+'_production']['old']) + parseInt(val.quantity);
+                productionItemsObj[val.sub_item_id+'_production']['old'] = parseFloat(productionItemsObj[val.sub_item_id+'_production']['old']) + parseFloat(val.quantity);
               }
-              // productionItemsObj[val.sub_item_id+'_production'] = parseInt(productionItemsObj[val.sub_item_id+'_production']) + parseInt(val.quantity);
+              // productionItemsObj[val.sub_item_id+'_production'] = parseFloat(productionItemsObj[val.sub_item_id+'_production']) + parseFloat(val.quantity);
             } else {
               productionItemsObj[val.sub_item_id+'_production'] = {};
-              productionItemsObj[val.sub_item_id+'_production']['quantity'] = parseInt(val.quantity);
+              productionItemsObj[val.sub_item_id+'_production']['quantity'] = parseFloat(val.quantity);
               productionItemsObj[val.sub_item_id+'_production']['today'] = 0;
               productionItemsObj[val.sub_item_id+'_production']['old'] = 0;
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                productionItemsObj[val.sub_item_id+'_production']['today'] = parseInt(productionItemsObj[val.sub_item_id+'_production']['today']) + parseInt(val.quantity);
+                productionItemsObj[val.sub_item_id+'_production']['today'] = parseFloat(productionItemsObj[val.sub_item_id+'_production']['today']) + parseFloat(val.quantity);
               } else {
-                productionItemsObj[val.sub_item_id+'_production']['old'] = parseInt(productionItemsObj[val.sub_item_id+'_production']['old']) + parseInt(val.quantity);
+                productionItemsObj[val.sub_item_id+'_production']['old'] = parseFloat(productionItemsObj[val.sub_item_id+'_production']['old']) + parseFloat(val.quantity);
               }
             }            
           } else {
             if(productionItemsObj[val.item_id+'_production']){
-              productionItemsObj[val.item_id+'_production']['quantity'] = parseInt(productionItemsObj[val.item_id+'_production']['quantity']) + parseInt(val.quantity);
+              productionItemsObj[val.item_id+'_production']['quantity'] = parseFloat(productionItemsObj[val.item_id+'_production']['quantity']) + parseFloat(val.quantity);
               productionItemsObj[val.item_id+'_production'] = productionItemsObj[val.item_id+'_production'];
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                productionItemsObj[val.item_id+'_production']['today'] = parseInt(productionItemsObj[val.item_id+'_production']['today']) + parseInt(val.quantity);
+                productionItemsObj[val.item_id+'_production']['today'] = parseFloat(productionItemsObj[val.item_id+'_production']['today']) + parseFloat(val.quantity);
               } else {
-                productionItemsObj[val.item_id+'_production']['old'] = parseInt(productionItemsObj[val.item_id+'_production']['old']) + parseInt(val.quantity);
+                productionItemsObj[val.item_id+'_production']['old'] = parseFloat(productionItemsObj[val.item_id+'_production']['old']) + parseFloat(val.quantity);
               }
-              // productionItemsObj[val.item_id+'_production'] = parseInt(productionItemsObj[val.item_id+'_production']) + parseInt(val.quantity);
+              // productionItemsObj[val.item_id+'_production'] = parseFloat(productionItemsObj[val.item_id+'_production']) + parseFloat(val.quantity);
             } else {
               productionItemsObj[val.item_id+'_production'] = {};
-              productionItemsObj[val.item_id+'_production']['quantity'] = parseInt(val.quantity);
+              productionItemsObj[val.item_id+'_production']['quantity'] = parseFloat(val.quantity);
               productionItemsObj[val.item_id+'_production']['today'] = 0;
               productionItemsObj[val.item_id+'_production']['old'] = 0;
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                productionItemsObj[val.item_id+'_production']['today'] = parseInt(productionItemsObj[val.item_id+'_production']['today']) + parseInt(val.quantity);
+                productionItemsObj[val.item_id+'_production']['today'] = parseFloat(productionItemsObj[val.item_id+'_production']['today']) + parseFloat(val.quantity);
               } else {
-                productionItemsObj[val.item_id+'_production']['old'] = parseInt(productionItemsObj[val.item_id+'_production']['old']) + parseInt(val.quantity);
+                productionItemsObj[val.item_id+'_production']['old'] = parseFloat(productionItemsObj[val.item_id+'_production']['old']) + parseFloat(val.quantity);
               }
             }
           }
@@ -178,52 +178,52 @@ router.get('/stock_movement', PermissionModule.Permission('view', moduleSlug,  e
         results.map(function(val, index){
           if(val.sub_item_id){
             if(invoiceItemsObj[val.sub_item_id+'_'+val.type]){
-              invoiceItemsObj[val.sub_item_id+'_'+val.type]['quantity'] = parseInt(invoiceItemsObj[val.sub_item_id+'_'+val.type]['quantity']) + parseInt(val.quantity);
+              invoiceItemsObj[val.sub_item_id+'_'+val.type]['quantity'] = parseFloat(invoiceItemsObj[val.sub_item_id+'_'+val.type]['quantity']) + parseFloat(val.quantity);
               invoiceItemsObj[val.sub_item_id+'_'+val.type] = invoiceItemsObj[val.sub_item_id+'_'+val.type];
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                invoiceItemsObj[val.sub_item_id+'_'+val.type]['today'] = parseInt(invoiceItemsObj[val.sub_item_id+'_'+val.type]['today']) + parseInt(val.quantity);
+                invoiceItemsObj[val.sub_item_id+'_'+val.type]['today'] = parseFloat(invoiceItemsObj[val.sub_item_id+'_'+val.type]['today']) + parseFloat(val.quantity);
               } else {
-                invoiceItemsObj[val.sub_item_id+'_'+val.type]['old'] = parseInt(invoiceItemsObj[val.sub_item_id+'_'+val.type]['old']) + parseInt(val.quantity);
+                invoiceItemsObj[val.sub_item_id+'_'+val.type]['old'] = parseFloat(invoiceItemsObj[val.sub_item_id+'_'+val.type]['old']) + parseFloat(val.quantity);
               }
-              // invoiceItemsObj[val.sub_item_id+'_'+val.type] = parseInt(invoiceItemsObj[val.sub_item_id+'_'+val.type]) + parseInt(val.quantity);
+              // invoiceItemsObj[val.sub_item_id+'_'+val.type] = parseFloat(invoiceItemsObj[val.sub_item_id+'_'+val.type]) + parseFloat(val.quantity);
             } else {
               // invoiceItemsObj[val.sub_item_id+'_'+val.type] = {};
-              // invoiceItemsObj[val.sub_item_id+'_'+val.type]['quantity'] = parseInt(val.quantity);
+              // invoiceItemsObj[val.sub_item_id+'_'+val.type]['quantity'] = parseFloat(val.quantity);
               invoiceItemsObj[val.sub_item_id+'_'+val.type] = {};
-              invoiceItemsObj[val.sub_item_id+'_'+val.type]['quantity'] = parseInt(val.quantity);
+              invoiceItemsObj[val.sub_item_id+'_'+val.type]['quantity'] = parseFloat(val.quantity);
               invoiceItemsObj[val.sub_item_id+'_'+val.type]['today'] = 0;
               invoiceItemsObj[val.sub_item_id+'_'+val.type]['old'] = 0;
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                invoiceItemsObj[val.sub_item_id+'_'+val.type]['today'] = parseInt(invoiceItemsObj[val.sub_item_id+'_'+val.type]['today']) + parseInt(val.quantity);
+                invoiceItemsObj[val.sub_item_id+'_'+val.type]['today'] = parseFloat(invoiceItemsObj[val.sub_item_id+'_'+val.type]['today']) + parseFloat(val.quantity);
               } else {
-                invoiceItemsObj[val.sub_item_id+'_'+val.type]['old'] = parseInt(invoiceItemsObj[val.sub_item_id+'_'+val.type]['old']) + parseInt(val.quantity);
+                invoiceItemsObj[val.sub_item_id+'_'+val.type]['old'] = parseFloat(invoiceItemsObj[val.sub_item_id+'_'+val.type]['old']) + parseFloat(val.quantity);
               }
             }
           } else {
             if(invoiceItemsObj[val.item_id+'_'+val.type]) {
-              invoiceItemsObj[val.item_id+'_'+val.type]['quantity'] = parseInt(invoiceItemsObj[val.item_id+'_'+val.type]['quantity']) + parseInt(val.quantity);
+              invoiceItemsObj[val.item_id+'_'+val.type]['quantity'] = parseFloat(invoiceItemsObj[val.item_id+'_'+val.type]['quantity']) + parseFloat(val.quantity);
               invoiceItemsObj[val.item_id+'_'+val.type] = invoiceItemsObj[val.item_id+'_'+val.type];
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                invoiceItemsObj[val.item_id+'_'+val.type]['today'] = parseInt(invoiceItemsObj[val.item_id+'_'+val.type]['today']) + parseInt(val.quantity);
+                invoiceItemsObj[val.item_id+'_'+val.type]['today'] = parseFloat(invoiceItemsObj[val.item_id+'_'+val.type]['today']) + parseFloat(val.quantity);
               } else {
-                invoiceItemsObj[val.item_id+'_'+val.type]['old'] = parseInt(invoiceItemsObj[val.item_id+'_'+val.type]['old']) + parseInt(val.quantity);
+                invoiceItemsObj[val.item_id+'_'+val.type]['old'] = parseFloat(invoiceItemsObj[val.item_id+'_'+val.type]['old']) + parseFloat(val.quantity);
               }
-              // invoiceItemsObj[val.item_id+'_'+val.type] = parseInt(invoiceItemsObj[val.item_id+'_'+val.type]) + parseInt(val.quantity);
+              // invoiceItemsObj[val.item_id+'_'+val.type] = parseFloat(invoiceItemsObj[val.item_id+'_'+val.type]) + parseFloat(val.quantity);
             } else {
               // invoiceItemsObj[val.item_id+'_'+val.type] = {};
-              // invoiceItemsObj[val.item_id+'_'+val.type]['quantity'] = parseInt(val.quantity);
+              // invoiceItemsObj[val.item_id+'_'+val.type]['quantity'] = parseFloat(val.quantity);
               invoiceItemsObj[val.item_id+'_'+val.type] = {};
-              invoiceItemsObj[val.item_id+'_'+val.type]['quantity'] = parseInt(val.quantity);
+              invoiceItemsObj[val.item_id+'_'+val.type]['quantity'] = parseFloat(val.quantity);
               invoiceItemsObj[val.item_id+'_'+val.type]['today'] = 0;
               invoiceItemsObj[val.item_id+'_'+val.type]['old'] = 0;
 
               if(helper.compareDate(helper.curDate('DD-MM-YYYY'), helper.changeDateFormate(val.createdAt.trim(), "YYYY-MM-DD", "DD-MM-YYYY"), 'DD-MM-YYYY') === 0){
-                invoiceItemsObj[val.item_id+'_'+val.type]['today'] = parseInt(invoiceItemsObj[val.item_id+'_'+val.type]['today']) + parseInt(val.quantity);
+                invoiceItemsObj[val.item_id+'_'+val.type]['today'] = parseFloat(invoiceItemsObj[val.item_id+'_'+val.type]['today']) + parseFloat(val.quantity);
               } else {
-                invoiceItemsObj[val.item_id+'_'+val.type]['old'] = parseInt(invoiceItemsObj[val.item_id+'_'+val.type]['old']) + parseInt(val.quantity);
+                invoiceItemsObj[val.item_id+'_'+val.type]['old'] = parseFloat(invoiceItemsObj[val.item_id+'_'+val.type]['old']) + parseFloat(val.quantity);
               }
             }            
           }
@@ -266,10 +266,10 @@ router.get('/stock_movement', PermissionModule.Permission('view', moduleSlug,  e
       
 
       if(invoiceItemsObj[key]){
-        tmpObj.sale = parseInt(invoiceItemsObj[key]['quantity']);
-        tmpObj.oldOut = parseInt(invoiceItemsObj[key]['old']);
-        tmpObj.todayOut = parseInt(invoiceItemsObj[key]['today']);
-        tmpObj.remaining = parseInt(value.quantity) - parseInt(invoiceItemsObj[key]['quantity']);
+        tmpObj.sale = parseFloat(invoiceItemsObj[key]['quantity']);
+        tmpObj.oldOut = parseFloat(invoiceItemsObj[key]['old']);
+        tmpObj.todayOut = parseFloat(invoiceItemsObj[key]['today']);
+        tmpObj.remaining = parseFloat(value.quantity) - parseFloat(invoiceItemsObj[key]['quantity']);
       } else {
         tmpObj.sale = 0;
         tmpObj.oldOut = 0;
